Rewrite project to MVC
Rewrite project to MVC pattern as a part of task
