+++ iOS-Settings-App/Resources/Base.lproj/LaunchScreen.storyboard
