+++ iOS-Settings-App/Resources/Info.plist
